+++ test/testProject/filters/filterAsMix.js
@@ -1,0 +1,12 @@
+function FilterAsMix() { }
+
+FilterAsMix.prototype = {
+	
+	executePre : function(event, context, svContext) {
+		/*eslint no-unused-vars: ["error", { "args": "none" }]*/
+		console.log("Pre");
+		svContext.filterChainCallback.success();
+	}
+};
+
+module.exports = FilterAsMix;