+++ test/testProject/sessionHandlers/sessionHandlerAsExports.js
@@ -1,0 +1,7 @@
+module.exports = {
+	
+	sessionStart : function(event, context, svContext) {
+		/*eslint no-unused-vars: ["error", { "args": "none" }]*/
+		console.log('Start');
+	}
+};