+++ test/testProject/sessionHandlers/sessionHandlerAsObject.js
@@ -1,0 +1,13 @@
+function ObjectSessionHandler() {}
+
+ObjectSessionHandler.prototype.sessionStart = function(event, context, svContext) {
+	/*eslint no-unused-vars: ["error", { "args": "none" }]*/
+	console.log('Start');
+};
+
+ObjectSessionHandler.prototype.sessionEnd = function(event, context, svContext) {
+	/*eslint no-unused-vars: ["error", { "args": "none" }]*/
+	console.log('End');
+};
+
+module.exports = ObjectSessionHandler;